feat: refactor auditlog to component
