+++ modules/front-end/src/app/core/components/audit-log/audit-log.component.html
@@ -1,0 +1,6 @@
+<div class="audit-log">
+  <div class="audit-log-title">{{titlePrefix}} <a (click)="goToTargetPage()">{{name}}</a></div>
+  <div class="audit-log-createdat">{{ auditLog.createdAt | date: "YYYY-MM-dd HH:mm"}}</div>
+  <div class="audit-log-comment">{{auditLog.comment}}</div>
+  <div class="change-list" *ngIf="shouldShowChangeList" [innerHtml]="htmlDiff | safeHtml"></div>
+</div>
