+++ modules/front-end/src/app/features/safe/audit-logs/index/index.component.less
@@ -29,31 +29,6 @@ section.body-container {
       margin-top: 15px;
       border-radius: 10px;
       padding: 0 10px;
-
-      .audit-log {
-        display: flex;
-        flex-direction: column;
-        gap:5px;
-        padding: 10px;
-
-        .audit-log-title {
-          font-weight: 600;
-          font-size: 15px;
-        }
-
-        .audit-log-createdat {
-          color: @grey50-color;
-        }
-
-        .audit-log-comment {
-          color: @grey50-color;
-        }
-
-        .change-list {
-          border-left: 2px solid @grey40-color;
-          padding: 0 12px;
-        }
-      }
     }
   }
